NMT dynamic memtags

+++ src/hotspot/share/nmt/mallocHeader.cpp
@@ -33,8 +33,8 @@
 #include "utilities/ostream.hpp"
 
 // The malloc header, as well as the coming VMATree implementation, rely on MemTag
-// fitting into eight bits.
-STATIC_ASSERT(sizeof(MemTag) == sizeof(uint8_t));
+// fitting into 32 bits.
+STATIC_ASSERT(sizeof(MemTag) == sizeof(uint32_t));
 
 void MallocHeader::print_block_on_error(outputStream* st, address bad_address) const {
   assert(bad_address >= (address)this, "sanity");
